looks like everything's fine, but not exactly everything.
When I tried to register for existing email application failed

Expect that - all good B-)

add as default two users to db, and some todos for default users. I think that's all

+++ src/main/java/net/enjoy/springboot/registrationlogin/entity/ToDo.java
@@ -2,8 +2,6 @@ package net.enjoy.springboot.registrationlogin.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.Size;
-import jakarta.websocket.server.PathParam;
-import jdk.jfr.DataAmount;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -19,7 +17,7 @@ import java.time.LocalDate;
 public class ToDo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String username;
 
     @Size(min=5, message="Enter atleast 5 characters")
