+++ src/main/java/net/enjoy/springboot/registrationlogin/repository/ToDoRepository.java
@@ -1,11 +1,12 @@
 package net.enjoy.springboot.registrationlogin.repository;
 
+import net.enjoy.springboot.registrationlogin.entity.ToDo;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface ToDoRepository extends JpaRepository <ToDo, Integer> {
+public interface ToDoRepository extends JpaRepository <ToDo, Long> {
     List<ToDo> findByUsername(String username);
 }
